t_fl: added module

+++ src/utils/game/game_utils.h
@@ -6,7 +6,7 @@
 /*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 08:12:53 by mmarinel          #+#    #+#             */
-/*   Updated: 2022/11/06 19:08:55 by mmarinel         ###   ########.fr       */
+/*   Updated: 2022/11/12 10:27:43 by mmarinel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,14 +23,14 @@
 # include <stddef.h>
 
 //*		geometry
-t_2d_point	ft_rotate(t_2d_point vector, double ray_angle);
+t_2d_point	ft_rotate(t_2d_point vector, float ray_angle);
 
 //*		vecotr utils
-t_2d_point	ft_get_new_2dpt(double x, double y);
+t_2d_point	ft_get_new_2dpt(float x, float y);
 t_2d_matrix	ft_get_new_2dmtrx(t_2d_point col_1, t_2d_point col_2);
 
 //*		math utils
-double		ft_dot_product(t_2d_point v, t_2d_point w);
+float		ft_dot_product(t_2d_point v, t_2d_point w);
 t_2d_point	ft_matrix_multiplication(t_2d_matrix matrix, t_2d_point vector);
 
 //*		utils
