+++ src/utils/game/vector.c
@@ -6,13 +6,13 @@
 /*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/06 17:50:45 by mmarinel          #+#    #+#             */
-/*   Updated: 2022/11/10 12:00:18 by mmarinel         ###   ########.fr       */
+/*   Updated: 2022/11/12 10:28:07 by mmarinel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "game_utils.h"
 
-t_2d_point	ft_get_new_2dpt(double x, double y)
+t_2d_point	ft_get_new_2dpt(float x, float y)
 {
 	t_2d_point	pt;
